% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc2num.R
\name{abc2num}
\alias{abc2num}
\title{Transforms a vector of strings to a list of numeric vectors where "a" corresponds to 1, "b" corresponds to 2 and so on. If 'x' has length=1 a vector is returned.}
\usage{
abc2num(x, char = letters, ignorecase = FALSE)
}
\value{

}
\description{
Transforms a vector of strings to a list of numeric vectors where "a" corresponds to 1, "b" corresponds to 2 and so on. If 'x' has length=1 a vector is returned.
}
\examples{
\dontrun{}

}
