% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addScoreVideo_WCh15.R
\name{addScoreVideo_WCh15}
\alias{addScoreVideo_WCh15}
\title{***DECRIPTION MISSING***}
\usage{
addScoreVideo_WCh15(t = 0, T = 60, s = cbind(0, 0), S = NULL,
  w = 1000, h = 1000, margin = c(2, 2), dir = "~/Pictures",
  filename = "image",
  logo = "~/Documents/Diverse/Bordhockey/Debatt2000/E09_VM15/Yngve Aasheim-Alexey Zaharov best av 7/VMStigabrett.png",
  fps = 25, fadet = 0.5, pos = c(100, 900, 250), names = c("Aasheim",
  "Nilsson"), cex = 1.5, family = "serif", col = "navyblue", font = 4,
  cores = 1)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
