% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_oblate_spheroid.R
\name{area_oblate_spheroid}
\alias{area_oblate_spheroid}
\title{***DECRIPTION MISSING***}
\usage{
area_oblate_spheroid(axes, exact = TRUE, p = 1.6075)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
