% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowsl.R
\name{arrowsl}
\alias{arrowsl}
\title{***DECRIPTION MISSING***}
\usage{
arrowsl(x0, y0, x1 = x0, y1 = y0, l = 1, length = 0.25, angle = 30,
  code = 2, col = par("fg"), lty = par("lty"), lwd = par("lwd"), ...)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
