% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bingo.R
\name{bingo}
\alias{bingo}
\title{Draws 'n' integers between 'a' and 'b'.}
\usage{
bingo(n = 10, a = 1, b = 100, timelag = 10, N = 200, sleep = TRUE,
  intro = TRUE, Nintro = 50, players = NULL, text = NULL, maxw = 20,
  startsleep = 0.001)
}
\value{

}
\description{
Draws 'n' integers between 'a' and 'b'.
}
\examples{
\dontrun{}

}
