% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BL_Table.R
\name{BL_Table}
\alias{BL_Table}
\title{***DECRIPTION MISSING***}
\usage{
BL_Table(dir,
  dirRnk = "~/Documents/Diverse/Bordhockey/FantasyNBA/Rankings_2014",
  series = list("Bergensl", c("Bergen C", "KM Berge")), season = NULL,
  league = "BL", hjemas = FALSE)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
