% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTeams2dat.R
\name{convertTeams2dat}
\alias{convertTeams2dat}
\title{Reads all teams in the given directory and writes to a table to be read more quickly later.}
\usage{
convertTeams2dat(dir)
}
\value{

}
\description{
Reads all teams in the given directory and writes to a table to be read more quickly later.
}
\examples{
\dontrun{}

}
