% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropImage.R
\name{cropImage}
\alias{cropImage}
\title{***DECRIPTION MISSING***}
\usage{
cropImage(im, margin = c(1, 1, 1, 1), thr = 1, crop = TRUE)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
