% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetaCor3d.R
\name{dbetaCor3d}
\alias{dbetaCor3d}
\title{***DECRIPTION MISSING***}
\usage{
dbetaCor3d(N = 100, cor = c(0.8, 0.4), lower = c(50, 50, 10),
  mode = c(150, 150, 30), upper = c(300, 300, 100), alpha = 2)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
