% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drayl.R
\name{drayl}
\alias{drayl}
\title{Returns probability density values of the Rayleigh distribution.}
\usage{
drayl(x, sigma)
}
\value{

}
\description{
Returns probability density values of the Rayleigh distribution.
}
\examples{
\dontrun{}

}
