% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.loadbalance_compact_school.R
\name{echoIBM.loadbalance_compact_school}
\alias{echoIBM.loadbalance_compact_school}
\title{Returns a vector of time steps that balance load on each core given the compactly specified school information.}
\usage{
echoIBM.loadbalance_compact_school(tlist, school, cores, runs = 10000)
}
\arguments{
\item{tlist}{is a list of the numbers of the pings to be simulated, where each element of the list corresponds to one compactlyu specified school.}

\item{school}{is a list of compactly specified dynamic school information.}

\item{cores}{is an integer specifying the number of cores to run the simulations over in parallel (should be lower than the number of cores in the computer).}

\item{runs}{is n integer giving the number of runs used to generate the optimal order.}
}
\value{

}
\description{
Returns a vector of time steps that balance load on each core given the compactly specified school information.
}
\examples{
\dontrun{}

}
