% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.read.workLSSS2.R
\name{echoIBM.read.workLSSS2}
\alias{echoIBM.read.workLSSS2}
\title{Reads an LSSS work file.}
\usage{
echoIBM.read.workLSSS2(f, lenb = 1324, numb = 500, flatten = TRUE)
}
\arguments{
\item{f}{is the file to read.}
}
\value{

}
\description{
Reads an LSSS work file.
}
\examples{
\dontrun{}

}
