% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.workLSSS2sgsc.event.R
\name{echoIBM.workLSSS2sgsc.event}
\alias{echoIBM.workLSSS2sgsc.event}
\title{Locates the path of an event specified by numeric or sting for both 'event' and 'cruise' (used in read.event()).}
\usage{
echoIBM.workLSSS2sgsc.event(event, segdir = NULL, cruise = 2009116,
  t = "all", esnm = "MS70", dir.data = NULL, code = NULL, ...)
}
\arguments{
\item{event}{is the identifier of the event in which the .work files are located.}

\item{segdir}{is an optional directory in which to put the segmentation files, defaulted to .}
}
\value{

}
\description{
Locates the path of an event specified by numeric or sting for both 'event' and 'cruise' (used in read.event()).
}
\examples{
\dontrun{}

}
