% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM_rename.R
\name{echoIBM_rename}
\alias{echoIBM_rename}
\title{Renames all files in the directory 'case' which contain the case name given as the last directory in the path of 'case', to the new case name given by newname.}
\usage{
echoIBM_rename(case, newname)
}
\arguments{
\item{case}{is the path to the directory holding the simulation case, ending with the case name after the last "/".}

\item{newname}{is the new case name.}
}
\value{

}
\description{
Renames all files in the directory 'case' which contain the case name given as the last directory in the path of 'case', to the new case name given by newname.
}
\examples{
\dontrun{}

}
