% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM_SX90_BlindZone.R
\name{echoIBM_SX90_BlindZone}
\alias{echoIBM_SX90_BlindZone}
\title{***DECRIPTION MISSING***}
\usage{
echoIBM_SX90_BlindZone(events = NULL, dir = NULL, project = NULL,
  eventspec = NULL, esnm = NULL, starttime = ftim2utim(20150101),
  seed = 1, lon0 = 5, lat0 = 70, beamPatternFish = FALSE,
  pingduration = 5, speedOfVessel = 10, duration = 12 * 3600,
  schooDist = 1000, maxRange = 450, maxSize = 100, maxHeight = 10,
  minDepth = -150, sonar_tilt = 3, sonar_plsl = 0.006, cores = 1,
  t = "all", pre = TRUE, schools = list(), CTD = FALSE,
  addnoise = TRUE, rand.sel = list(EK60 = 1, SU90 = 10), ...)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
