% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ffmpeg}
\alias{ffmpeg}
\alias{cutVideo}
\title{Run an ffmpeg command.}
\usage{
ffmpeg(f, fout = NULL, par = NULL, pre = NULL, y = TRUE, run = TRUE,
  dir = NULL, ...)

cutVideo(f, ss = "00:00:00", t = NULL, to = NULL, fout = NULL,
  par = NULL, ...)
}
\arguments{
\item{f}{The file path(s).}

\item{par}{Parameters, given as c(parameterName, parameterValue), such as c("-acodec", "copy", "-to", "00.:00:13")}

\item{pre}{Pre-parameters, given as c(parameterName, parameterValue), such as c("-ss", "00:00:00"),}

\item{y}{Logical: If TRUE overwrite.}

\item{run}{Logical: If TRUE run the command.}

\item{dir}{Directory in which to place the output files if 'fout' is not given.}

\item{...}{Used in fmod().}

\item{f}{The output file path(s).}

\item{f}{The file path(s).}

\item{f}{The output file path(s).}

\item{par}{Parameters, given as c(parameterName, parameterValue), such as c("-acodec", "copy", "-to", "00.:00:13")}

\item{pre}{Pre-parameters, given as c(parameterName, parameterValue), such as c("-ss", "00:00:00"),}

\item{y}{Logical: If TRUE overwrite.}

\item{run}{Logical: If TRUE run the command.}

\item{dir}{Directory in which to place the output files if 'fout' is not given.}

\item{...}{Used in fmod().}
}
\description{
Run an ffmpeg command on the files 'f' with output files 'fout', using parameters 'par' and pre-parameters 'pre'.

Cut video using ffmpeg.
}
