% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdBbs.R
\name{fitdBbs}
\alias{fitdBbs}
\title{Deduces a formula for the dBbs using Xsbr, and DISCARD PINGS WHERE THE AREA IS NOT CROSSING.}
\usage{

  fitdBbs(event = "/Users/arnejh/Data/echoIBM/SX90_biomassEstimation/Events/SX90_biomassEstimation_event_one_school_at_the_time_directional_fish_realisticNoise/SX90/tsd",
  dBan = 8, dBbs = 1:20)
}
\arguments{
\item{event}{is the event holding the segmentation data.}

\item{dBbs}{are the dBbs values used in the segmentation data.}

\item{segpar}{is the file number of the initial segmentation (defined from the background level).}
}
\value{

}
\description{
Deduces a formula for the dBbs using Xsbr, and DISCARD PINGS WHERE THE AREA IS NOT CROSSING.
}
\examples{
\dontrun{}

}
