% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.list.R
\name{flatten.list}
\alias{flatten.list}
\title{Flattens lists so that all R-objects to the desired depth in the input list 'x' are placed in separate positions in the output list. Differs from unlist() by the fact that unlist() flattens an array to single elements.}
\usage{
flatten.list(x, depth = Inf)
}
\value{

}
\description{
Flattens lists so that all R-objects to the desired depth in the input list 'x' are placed in separate positions in the output list. Differs from unlist() by the fact that unlist() flattens an array to single elements.
}
\examples{
\dontrun{}

}
