% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{fmod}
\alias{fmod}
\title{Modify file name, adding text before file extension, changing file extension or changing directory.}
\usage{
fmod(f, add, ext = NULL, dir = NULL, ...)
}
\arguments{
\item{f}{The file path(s).}

\item{add}{The string to add to each path.}

\item{ext}{The new file extension.}

\item{dir}{The new directory.}

\item{...}{}
}
\description{
Modify file name, adding text before file extension, changing file extension or changing directory.
}
