% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.test.exp.R
\name{fs.test.exp}
\alias{fs.test.exp}
\title{Tests for exponentiality using the Kolmogorov-Smirnov test for exponential distribution when mean is unknown and must be estimated from the data. Critical values are from Monte Carlo simulations done by Finkelstein and Schafer 1972, but resimulated by Holmin (see "/Applications/echoIBM/Diverse/Finkelstein.Schafer.S.R").}
\usage{
fs.test.exp(x, p = 0.1)
}
\arguments{
\item{x}{is a vector to be tested for exponentiality. Must have length 2, ..., 20, 25, 30, 100 or 1000.}

\item{p}{is the significance level, one of 0.2, 0.15, 0.1, 0.05, 0.01, 0.005, 0.001.}
}
\value{

}
\description{
Tests for exponentiality using the Kolmogorov-Smirnov test for exponential distribution when mean is unknown and must be estimated from the data. Critical values are from Monte Carlo simulations done by Finkelstein and Schafer 1972, but resimulated by Holmin (see "/Applications/echoIBM/Diverse/Finkelstein.Schafer.S.R").
}
\examples{
\dontrun{}

}
