% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.filename.R
\name{get.filename}
\alias{get.filename}
\title{Extracts all but the file extensions from a vector of file names (discarding the dot by default).}
\usage{
get.filename(filelist, includedot = FALSE)
}
\arguments{
\item{filelist}{is a string vector of file names, from which the file extension is extracted.}
}
\value{

}
\description{
Extracts all but the file extensions from a vector of file names (discarding the dot by default).
}
\examples{
\dontrun{}

}
