% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.points_fantasyNBA.R
\name{get.points_fantasyNBA}
\alias{get.points_fantasyNBA}
\title{***DECRIPTION MISSING***}
\usage{
get.points_fantasyNBA(dir, pow = c(3, 3, 3, 3, 1, 1), w = c(32, 32, 31, 28,
  10, 6), t = "all", n = "all")
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
