% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getallplayers.R
\name{getallplayers}
\alias{getallplayers}
\title{Returns a data frame of all players in the FantasyNBA.}
\usage{
getallplayers()
}
\value{

}
\description{
Returns a data frame of all players in the FantasyNBA.
}
\examples{
\dontrun{}

}
