% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEvents.R
\name{getEvents}
\alias{getEvents}
\title{***DECRIPTION MISSING***}
\usage{
getEvents(eventspec, dir = "/Users/arnejh/Data/echoIBM",
  project = "SX90_BlindZone", esnm = c("SU90", "EK60"), new = TRUE)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
