% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getprice_testseason.R
\name{getprice_testseason}
\alias{getprice_testseason}
\title{Returns the dataframe of prices for all gameweeks, using the formula 2 + WR * w + (T + (P-G)*0.5) * s.}
\usage{
getprice_testseason(date = NULL, sort = FALSE, volume = "all")
}
\value{

}
\description{
Returns the dataframe of prices for all gameweeks, using the formula 2 + WR * w + (T + (P-G)*0.5) * s.
}
\examples{
\dontrun{}

}
