% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpriceWR.R
\name{getpriceWR}
\alias{getpriceWR}
\title{Reads the data in copy-paste files from the world ranking in table hockey.}
\usage{
getpriceWR(file = NULL, date = NULL, par = c(0.01, 2/3))
}
\value{

}
\description{
Reads the data in copy-paste files from the world ranking in table hockey.
}
\examples{
\dontrun{}

}
