% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWR.R
\name{getWR}
\alias{getWR}
\title{***DECRIPTION MISSING***}
\usage{
getWR(players = c("all", "nor+", "nor"), date = NULL, file = NULL)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
