% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWR_2014.R
\name{getWR_2014}
\alias{getWR_2014}
\title{***DECRIPTION MISSING***}
\usage{
getWR_2014(players = c("all", "nor+", "nor"), date = NULL, file = NULL)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
