% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIB.MS70.R
\name{HIB.MS70}
\alias{HIB.MS70}
\title{Processing and writing datasets by the HIB-algorithm. See 'The HIB-algorithm.pdf' for documentation and read.event() for importing HIBs. The processing needs to be performed on data that is not compensated for absorption and attenuation through the TVG function.}
\usage{
HIB.MS70(cruise = 2009116, event = 1, t = "all", C = 1e-06, nu = 1,
  sep = 101, turns = 10, near_range = 50, TVG = FALSE,
  dir.data = NULL, tofile = TRUE)
}
\value{

}
\description{
Processing and writing datasets by the HIB-algorithm. See 'The HIB-algorithm.pdf' for documentation and read.event() for importing HIBs. The processing needs to be performed on data that is not compensated for absorption and attenuation through the TVG function.
}
\examples{
\dontrun{}

}
