% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histexp.R
\name{histexp}
\alias{histexp}
\title{Plots a histogram of 'x' added a line of the maximum likelihood fitted exponential distribution to 'x'.}
\usage{
histexp(x, rate = NULL, cut = NULL, type = "l", scale = 1,
  breaks = "Sturges", coll = "red", lengthx = 1000, lwd = 1.5,
  add = FALSE, plot.all = TRUE, pres = cut * 1e-04, max.iter = 100,
  plot = TRUE, ...)
}
\value{

}
\description{
Plots a histogram of 'x' added a line of the maximum likelihood fitted exponential distribution to 'x'.
}
\examples{
\dontrun{}

}
