% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histgamma.R
\name{histgamma}
\alias{histgamma}
\title{Plots a histogram of 'x' added a line of the maximum likelihood fitted gamma distribution to 'x'.}
\usage{
histgamma(x, breaks = "Sturges", coll = "red", lengthx = 1000,
  lwd = 1.5, add = FALSE, plot = TRUE, ...)
}
\value{

}
\description{
Plots a histogram of 'x' added a line of the maximum likelihood fitted gamma distribution to 'x'.
}
\examples{
\dontrun{}

}
