% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histnorm.R
\name{histnorm}
\alias{histnorm}
\title{Plots a histogram of 'x' added a line of the maximum likelihood fitted Gaussian distribution to 'x'.}
\usage{
histnorm(x, mu = NULL, sigma = NULL, breaks = "Sturges", coll = "red",
  lengthx = 1000, lwd = 1.5, add = FALSE, plot = TRUE, ...)
}
\value{

}
\description{
Plots a histogram of 'x' added a line of the maximum likelihood fitted Gaussian distribution to 'x'.
}
\examples{
\dontrun{}

}
