% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histpois.R
\name{histpois}
\alias{histpois}
\title{Plots a histogram of 'x' added a line of the maximum likelihood fitted Poisson distribution to 'x'.}
\usage{
histpois(x, mu = NULL, breaks = "Sturges", coll = "red", lengthx = 1000,
  lwd = 1.5, add = FALSE, plot = TRUE, ...)
}
\value{

}
\description{
Plots a histogram of 'x' added a line of the maximum likelihood fitted Poisson distribution to 'x'.
}
\examples{
\dontrun{}

}
