% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histrayl.R
\name{histrayl}
\alias{histrayl}
\title{Plots a histogram of 'x' added a line of the maximum likelihood fitted Rayleigh distribution to 'x'.}
\usage{
histrayl(x, sigma = NULL, breaks = "Sturges", coll = "red",
  lengthx = 1000, lwd = 1.5, add = FALSE, plot = TRUE, ...)
}
\value{

}
\description{
Plots a histogram of 'x' added a line of the maximum likelihood fitted Rayleigh distribution to 'x'.
}
\examples{
\dontrun{}

}
