% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histweibull.R
\name{histweibull}
\alias{histweibull}
\title{Plots a histogram of 'x' added a line of the maximum likelihood fitted Weibull distribution to 'x'.}
\usage{
histweibull(x, breaks = "Sturges", coll = "red", lengthx = 1000,
  lwd = 1.5, add = FALSE, plot = TRUE, ...)
}
\value{

}
\description{
Plots a histogram of 'x' added a line of the maximum likelihood fitted Weibull distribution to 'x'.
}
\examples{
\dontrun{}

}
