% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2str.R
\name{int2str}
\alias{int2str}
\title{Creates a vector of strings holding the integers given by 'int' added zeros to the left so that all strings have length 'len'.}
\usage{
int2str(int, len = 0)
}
\value{

}
\description{
Creates a vector of strings holding the integers given by 'int' added zeros to the left so that all strings have length 'len'.
}
\examples{
\dontrun{}

}
