% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intseq.R
\name{intseq}
\alias{intseq}
\title{Transforms a number to a vector of single intergers. Commas returned as NA. Very slow function.}
\usage{
intseq(num)
}
\value{

}
\description{
Transforms a number to a vector of single intergers. Commas returned as NA. Very slow function.
}
\examples{
\dontrun{}

}
