% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.flat.R
\name{is.flat}
\alias{is.flat}
\title{Checks for equality of all elements in an array given the tolerance.}
\usage{
is.flat(x, tol = 0)
}
\value{

}
\description{
Checks for equality of all elements in an array given the tolerance.
}
\examples{
\dontrun{}

}
