% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iTunes.random.R
\name{iTunes.random}
\alias{iTunes.random}
\title{Creates a new random playlist from the existing playlist 'infile'. The playlist 'infile' must be created by using the "Archive-Library-Export playlist" feature ("Arkiv-Bibliotek-Eksporter spilleliste" in norwegian) in "simple text"/"plain text" mode ("Enkel tekst" in norwegian).}
\usage{
iTunes.random(infile = NULL, outfile = "Random_playlist.txt", size = 100,
  Rating = 1, Frequency = 1, Genre = list("all", 1), exclude = NULL,
  replace = FALSE, language = "norwegian")
}
\value{

}
\description{
Creates a new random playlist from the existing playlist 'infile'. The playlist 'infile' must be created by using the "Archive-Library-Export playlist" feature ("Arkiv-Bibliotek-Eksporter spilleliste" in norwegian) in "simple text"/"plain text" mode ("Enkel tekst" in norwegian).
}
\examples{
\dontrun{}

}
