% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernSmooth3dGauss.TSD.R
\name{kernSmooth3dGauss.TSD}
\alias{kernSmooth3dGauss.TSD}
\title{Applies a Gaussian 3-D kernel on MS70 sonar data given as a list 'data' with names according to the TSD-convension, and the bandwidths given in the three element vector 'h':}
\usage{
kernSmooth3dGauss.TSD(data, h = c(1, 1, 1), w = h * 3, ind = list())
}
\arguments{
\item{data}{is the list of data with names according to the TSD-convension.}

\item{h}{is the three element vector of bandwidths.}

\item{w}{is the limits of the kernel, outside which the kernel is 0.}
}
\value{

}
\description{
Applies a Gaussian 3-D kernel on MS70 sonar data given as a list 'data' with names according to the TSD-convension, and the bandwidths given in the three element vector 'h':
}
\examples{
\dontrun{}

}
