% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernSmoothGauss.R
\name{kernSmoothGauss}
\alias{kernSmoothGauss}
\title{Applies a Gaussian 3-D kernel on MS70 sonar data given as the coordinates 'x' and the data 'y', and the bandwidths given in the three element vector 'h':}
\usage{
kernSmoothGauss(x, y, h = 1, w = h * 3)
}
\arguments{
\item{x}{is the matrix of the coordinates.}

\item{y}{is the vector of the data, with dimension corresponding to the number of columns if 'x'.}

\item{h}{is the three element vector of bandwidths.}

\item{w}{is the limits of the kernel, outside which the kernel is 0.}
}
\value{

}
\description{
Applies a Gaussian 3-D kernel on MS70 sonar data given as the coordinates 'x' and the data 'y', and the bandwidths given in the three element vector 'h':
}
\examples{
\dontrun{}

}
