% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.min.R
\name{local.min}
\alias{local.min}
\title{Locates local minima of a vector 'y' (given the vector 'x' of equal length). Method based on the definition given in Calculus, Fourth Edition by Robert A. Adams: (local minima occur iff there exists a h>0 so that y(x*)<=y(x) at all points |x*-x|<h).}
\usage{
local.min(x, y = NULL, strict = TRUE, w = 1)
}
\value{

}
\description{
Locates local minima of a vector 'y' (given the vector 'x' of equal length). Method based on the definition given in Calculus, Fourth Edition by Robert A. Adams: (local minima occur iff there exists a h>0 so that y(x*)<=y(x) at all points |x*-x|<h).
}
\examples{
\dontrun{}

}
