% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.calsphere.MS70.R
\name{locate.calsphere.MS70}
\alias{locate.calsphere.MS70}
\title{Uses locate.calsphere.MS70.voxel() to locate the voxel containing the calibration sphere of MS70 calibration data, and the split beam angle information "angl" and "angt" to furhter pinpoint the position.}
\usage{
locate.calsphere.MS70(event, t, ind = list(-(1:30), NULL, NULL), K = 50,
  step = c(2, 2, 2), margin = c(3, 2, 2), firstpos = NULL,
  sizemargin = c(0.1, 10), write = FALSE,
  filename = "_Calibration_sphere_location.tsd")
}
\arguments{
\item{event}{is a string naming the calibration event.}

\item{t}{is either a numeric vector of time step indexes, or "all" which implies treating all time steps.}

\item{ind}{is a list of indexes subsetting the data using subset_TSD() (see subset_TSD() for details).}

\item{K}{is the number of the largest voxels to include in the clustering forming the basis of the funstion (see DESCRIPTION above).}

\item{step}{is a vector of three elements (radial, horizontal, vertical) representing the proximity around the old position in which the new position is accepted.}

\item{margin}{is a matrix of three elements (radial, horizontal, vertical) representing the distances separating clusters.}

\item{firstpos}{is a vector of three elements (radial, horizontal, vertical) representing the first voxel position of the calibration sphere.}

\item{sizemargin}{is a vector of two factors representing the interval on each size of 'oldsize' in which the new size must be contained. If sizemargin=[0.5,1.5] and 'size' of the relevant cluster is 10, the interval will be [5,15].}

\item{write}{TRUE if the data should be written to the event directory in a file named "eventname_Calibration_sphere_location.tsd", where 'eventname' is the name of the event, extracted from the string 'event'.}
}
\value{

}
\description{
Uses locate.calsphere.MS70.voxel() to locate the voxel containing the calibration sphere of MS70 calibration data, and the split beam angle information "angl" and "angt" to furhter pinpoint the position.
}
\examples{
\dontrun{}

}
