% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.calsphere.MS70.voxel.R
\name{locate.calsphere.MS70.voxel}
\alias{locate.calsphere.MS70.voxel}
\title{Locates the voxel containing the calibration sphere of calibration data from the MS70 sonar. The method used is a geometrical clustering of the 'K' voxels of highest intensity, folowed by a ranging of the clusters based on the sum of the intensities of each cluster. If 'oldpos' is missing the calibration sphere is assigned the voxel position of the largest cluster. If 'oldpos' is present, the location needs to be in the proximity of the old position, given by 'step'. Also the size of the chosen cluster needs to be close to the old size. If these conditions are not met, NULL is returned for the 'voxel'.}
\usage{
locate.calsphere.MS70.voxel(data, ind = list(-(1:30), NULL, NULL), K = 100,
  oldpos = NULL, oldsize = NULL, step = c(2, 2, 2), margin = c(3, 2, 2),
  sizemargin = c(0.5, 1.5))
}
\arguments{
\item{data}{is a list of data as returned from read.event(...,var=c("vbsc","voxels","beams")).}

\item{ind}{is a list of indexes subsetting the data using subset_TSD() (see subset_TSD() for details).}

\item{K}{is the number of the largest voxels to include in the clustering forming the basis of the funstion (see DESCRIPTION above).}

\item{oldpos}{is a vector of three elements (radial, horizontal, vertical) representing the old voxel position of the calibration sphere.}

\item{oldsize}{is a the sum of the voxels in old cluster.}

\item{step}{is a vector of three elements (radial, horizontal, vertical) representing the proximity around the old position in which the new position is accepted.}

\item{margin}{is a matrix of three elements (radial, horizontal, vertical) representing the distances separating clusters.}

\item{sizemargin}{is a vector of two factors representing the interval on each size of 'oldsize' in which the new size must be contained. If sizemargin=[0.5,1.5] and 'size' of the relevant cluster is 10, the interval will be [5,15].}
}
\value{

}
\description{
Locates the voxel containing the calibration sphere of calibration data from the MS70 sonar. The method used is a geometrical clustering of the 'K' voxels of highest intensity, folowed by a ranging of the clusters based on the sum of the intensities of each cluster. If 'oldpos' is missing the calibration sphere is assigned the voxel position of the largest cluster. If 'oldpos' is present, the location needs to be in the proximity of the old position, given by 'step'. Also the size of the chosen cluster needs to be close to the old size. If these conditions are not met, NULL is returned for the 'voxel'.
}
\examples{
\dontrun{}

}
