% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lremove.R
\name{lremove}
\alias{lremove}
\title{Removing elements equal to 'item' from a list.}
\usage{
lremove(x, item = NULL, robust = FALSE)
}
\value{

}
\description{
Removing elements equal to 'item' from a list.
}
\examples{
\dontrun{}

}
