% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwhich.R
\name{lwhich}
\alias{lwhich}
\title{Locating elements equal to 'item' in a list.}
\usage{
lwhich(x, item = NULL, fun = function(x) x, nomatch = NULL,
  robust = FALSE)
}
\value{

}
\description{
Locating elements equal to 'item' in a list.
}
\examples{
\dontrun{}

}
