% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateBeams.R
\name{manipulateBeams}
\alias{manipulateBeams}
\title{***DECRIPTION MISSING***}
\usage{
manipulateBeams(reference_beamsFile, outfile, bmmd = 0, numt = 1,
  utim = NULL, maxRange = NULL, tilt = NULL, plsl = NULL)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
