% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastermind.R
\name{mastermind}
\alias{mastermind}
\title{The classic Master Mind game.}
\usage{
mastermind(save.results = FALSE, save.dir = "Desktop")
}
\value{

}
\description{
The classic Master Mind game.
}
\examples{
\dontrun{}

}
