% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlerayl.R
\name{mlerayl}
\alias{mlerayl}
\title{Maximum likelihood estimate of a Rayleigh distributed variable.}
\usage{
mlerayl(x, na.rm = TRUE)
}
\value{

}
\description{
Maximum likelihood estimate of a Rayleigh distributed variable.
}
\examples{
\dontrun{}

}
