% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearly.duplicated.R
\name{nearly.duplicated}
\alias{nearly.duplicated}
\title{Determines which elements of an R object are nearly equal the previous element. Differs from duplicated by 2 means: (1) The use of nearly equality. (2) The comparison to the prevoius element and not all previous elements.}
\usage{
nearly.duplicated(x, tolerance = .Machine$double.eps^0.5)
}
\value{

}
\description{
Determines which elements of an R object are nearly equal the previous element. Differs from duplicated by 2 means: (1) The use of nearly equality. (2) The comparison to the prevoius element and not all previous elements.
}
\examples{
\dontrun{}

}
