% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearly.equal.R
\name{nearly.equal}
\alias{nearly.equal}
\title{Tests for near equality. To be used when == fails to equal values that are intended to be equal, due to machine inaccuracy. See info(nearly.equal). Related to all.equal() but returns a logical vector instead of only one logical.}
\usage{
nearly.equal(x, y, tolerance = .Machine$double.eps^0.5)
}
\value{

}
\description{
Tests for near equality. To be used when == fails to equal values that are intended to be equal, due to machine inaccuracy. See info(nearly.equal). Related to all.equal() but returns a logical vector instead of only one logical.
}
\examples{
\dontrun{}

}
