% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearly.rle.R
\name{nearly.rle}
\alias{nearly.rle}
\title{Computes the lengths and values of runs of nearly equal values in a vector.}
\usage{
nearly.rle(x, tolerance = .Machine$double.eps^0.5)
}
\value{

}
\description{
Computes the lengths and values of runs of nearly equal values in a vector.
}
\examples{
\dontrun{}

}
