% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearly.unique.R
\name{nearly.unique}
\alias{nearly.unique}
\title{Returns the unique values of the R object 'x', where nearly equal values are regarded as equal.}
\usage{
nearly.unique(x, tolerance = .Machine$double.eps^0.5)
}
\value{

}
\description{
Returns the unique values of the R object 'x', where nearly equal values are regarded as equal.
}
\examples{
\dontrun{}

}
