% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2abc.R
\name{num2abc}
\alias{num2abc}
\title{Transforms a numeric vector to a list of corresponding character arrays. If input has length=1, a vector is returned.}
\usage{
num2abc(num, abc = letters[1:10])
}
\value{

}
\description{
Transforms a numeric vector to a list of corresponding character arrays. If input has length=1, a vector is returned.
}
\examples{
\dontrun{}

}
