% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perreg.R
\name{perreg}
\alias{perreg}
\title{Periodic parametric regression on a periodic response variable.}
\usage{
perreg(x, y = NULL, d = NULL, m = NULL)
}
\value{

}
\description{
Periodic parametric regression on a periodic response variable.
}
\examples{
\dontrun{}

}
