% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploteigen.R
\name{ploteigen}
\alias{ploteigen}
\title{Plots eigenvectors of lengths given by 'eigenvalues' in 2D or 3D.}
\usage{
ploteigen(eigenvectors, eigenvalues = NULL, add = FALSE, center = 0,
  square = TRUE, col = "black", ...)
}
\value{

}
\description{
Plots eigenvectors of lengths given by 'eigenvalues' in 2D or 3D.
}
\examples{
\dontrun{}

}
