% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotone.R
\name{plotone}
\alias{plotone}
\title{***DECRIPTION MISSING***}
\usage{
plotone(i, dir, imfiles, aspectRatio, sec, margin, fadet, logo, bg, pos,
  dimlogo, dimbg, score, matchScore, cex, family, col, font, brightness, ind,
  names, adj, T)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
