% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarySeg.R
\name{plotSummarySeg}
\alias{plotSummarySeg}
\title{***DECRIPTION MISSING***}
\usage{
plotSummarySeg(s, var = "tons", x = "anio", ylim = NULL, col = 1,
  linecol = 1, cex = 1, spar = NULL, legpos = "topright",
  smooth = TRUE, cex.leg = 1.3, ...)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
