% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2lines3d.R
\name{points2lines3d}
\alias{points2lines3d}
\title{***DECRIPTION MISSING***}
\usage{
points2lines3d(x, y = NULL, steps = 1, atP = 3)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
