% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2score.R
\name{points2score}
\alias{points2score}
\title{***DECRIPTION MISSING***}
\usage{
points2score(x, pow = pow, w = 20, maxp = 900, bronze = TRUE)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
