% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prayl.R
\name{prayl}
\alias{prayl}
\title{Returns cummulative probability density values of the Rayleigh distribution.}
\usage{
prayl(q, sigma)
}
\value{

}
\description{
Returns cummulative probability density values of the Rayleigh distribution.
}
\examples{
\dontrun{}

}
