% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printt.R
\name{printt}
\alias{printt}
\title{Prints the text of 'expr' prior to the printing of 'expr'.}
\usage{
printt(expr = NULL, ...)
}
\arguments{
\item{expr}{is the expression to print.}

\item{...}{variables passed on to print().}
}
\value{

}
\description{
Prints the text of 'expr' prior to the printing of 'expr'.
}
\examples{
\dontrun{}

}
