% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_Bergensliga1.R
\name{process_Bergensliga1}
\alias{process_Bergensliga1}
\title{***DECRIPTION MISSING***}
\usage{
process_Bergensliga1(season = "H2011", p, pKM)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
