% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_Bergensligaen.R
\name{process_Bergensligaen}
\alias{process_Bergensligaen}
\title{***DECRIPTION MISSING***}
\usage{
process_Bergensligaen(seasons = c("H2011", "V2012"), type = c("points",
  "GSB"), p = c("Holmin", "Old", "FIS", "Biathlon"), v = 13)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
