% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrayl.R
\name{qrayl}
\alias{qrayl}
\title{Returns quantile values of the Rayleigh distribution.}
\usage{
qrayl(p, sigma)
}
\value{

}
\description{
Returns quantile values of the Rayleigh distribution.
}
\examples{
\dontrun{}

}
