% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank2bonus_old.R
\name{rank2bonus_old}
\alias{rank2bonus_old}
\title{***DECRIPTION MISSING***}
\usage{
rank2bonus_old(x, A, level = 4, bronze = TRUE)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
