% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank2wrPoints.R
\name{rank2wrPoints}
\alias{rank2wrPoints}
\title{Calculates WR points based on best scores.}
\usage{
rank2wrPoints(best, level = 5, bronze = TRUE)
}
\value{

}
\description{
Calculates WR points based on best scores.
}
\examples{
\dontrun{}

}
