% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank2wrPoints_2014.R
\name{rank2wrPoints_2014}
\alias{rank2wrPoints_2014}
\title{Calculates WR points based on best scores.}
\usage{
rank2wrPoints_2014(best, level = 5, bronze = TRUE, euro = FALSE)
}
\value{

}
\description{
Calculates WR points based on best scores.
}
\examples{
\dontrun{}

}
