% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbeta.R
\name{Rbeta}
\alias{Rbeta}
\title{Returns an array of randomly generated values from the beta distribution.}
\usage{
Rbeta(dim, shape1, shape2, ncp = 0, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the beta distribution.
}
\examples{
\dontrun{}

}
