% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbinom.R
\name{Rbinom}
\alias{Rbinom}
\title{Returns an array of randomly generated values from the binomial distribution.}
\usage{
Rbinom(dim, size, prob, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the binomial distribution.
}
\examples{
\dontrun{}

}
