% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcauchy.R
\name{Rcauchy}
\alias{Rcauchy}
\title{Returns an array of randomly generated values from the Cauchy distribution.}
\usage{
Rcauchy(dim, location = 0, scale = 1, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the Cauchy distribution.
}
\examples{
\dontrun{}

}
