% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rchisq.R
\name{Rchisq}
\alias{Rchisq}
\title{Returns an array of randomly generated values from the chi-squared distribution.}
\usage{
Rchisq(dim, df, ncp = 0, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the chi-squared distribution.
}
\examples{
\dontrun{}

}
