% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_power2sigmabs.R
\name{readEKRaw_power2sigmabs}
\alias{readEKRaw_power2sigmabs}
\title{Converts power to backscattering cross section from data read from a Simrad raw file.}
\usage{
readEKRaw_power2sigmabs(x)
}
\arguments{
\item{x}{is a list of the data.}
}
\value{

}
\description{
Converts power to backscattering cross section from data read from a Simrad raw file.
}
\examples{
\dontrun{}

}
