% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_power2sigmabs.TSD.R
\name{readEKRaw_power2sigmabs.TSD}
\alias{readEKRaw_power2sigmabs.TSD}
\title{Converts power to backscattering cross section from data in the TSD format.}
\usage{
readEKRaw_power2sigmabs.TSD(x, beams)
}
\arguments{
\item{x}{is a list of the data.}

\item{beams}{is a list of the beam configuration of the sonar or echosounder.}
}
\value{

}
\description{
Converts power to backscattering cross section from data in the TSD format.
}
\examples{
\dontrun{}

}
