% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_power2sv.R
\name{readEKRaw_power2sv}
\alias{readEKRaw_power2sv}
\title{Converts power to volume backscattering coefficient from data read from a Simrad raw file.}
\usage{
readEKRaw_power2sv(x)
}
\arguments{
\item{x}{is a list of the data.}
}
\value{

}
\description{
Converts power to volume backscattering coefficient from data read from a Simrad raw file.
}
\examples{
\dontrun{}

}
