% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfromWR.R
\name{readfromWR}
\alias{readfromWR}
\title{Reads the data in copy-paste files from the world ranking in table hockey.}
\usage{
readfromWR(x, nsign = 10)
}
\value{

}
\description{
Reads the data in copy-paste files from the world ranking in table hockey.
}
\examples{
\dontrun{}

}
