% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTeams.R
\name{readTeams}
\alias{readTeams}
\title{Reads all teams in the given directory using the quickly accessed file "AllTeams.dat".}
\usage{
readTeams(dir)
}
\value{

}
\description{
Reads all teams in the given directory using the quickly accessed file "AllTeams.dat".
}
\examples{
\dontrun{}

}
