% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroup_school.R
\name{regroup_school}
\alias{regroup_school}
\title{Writes the function name, description and variables explainaiton of the functions given by 'functions' to the available examples files given by 'examples'. The example file must have the structure given by "/Applications/R/example template".}
\usage{
regroup_school(pos = NULL, len = 1, adjust = 3)
}
\value{

}
\description{
Writes the function name, description and variables explainaiton of the functions given by 'functions' to the available examples files given by 'examples'. The example file must have the structure given by "/Applications/R/example template".
}
\examples{
\dontrun{}

}
