% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rexp.R
\name{Rexp}
\alias{Rexp}
\title{Returns an array of randomly generated values from the exponential distribution.}
\usage{
Rexp(dim, rate = 1, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the exponential distribution.
}
\examples{
\dontrun{}

}
