% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rf.R
\name{Rf}
\alias{Rf}
\title{Returns an array of randomly generated values from the F distribution.}
\usage{
Rf(dim, df1, df2, ncp, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the F distribution.
}
\examples{
\dontrun{}

}
