% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgamma.R
\name{Rgamma}
\alias{Rgamma}
\title{Returns an array of randomly generated values from the gamma distribution.}
\usage{
Rgamma(dim, shape, rate = 1, scale = 1/rate, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the gamma distribution.
}
\examples{
\dontrun{}

}
