% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgeom.R
\name{Rgeom}
\alias{Rgeom}
\title{Returns an array of randomly generated values from the geometric distribution.}
\usage{
Rgeom(dim, prob, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the geometric distribution.
}
\examples{
\dontrun{}

}
