% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhyper.R
\name{Rhyper}
\alias{Rhyper}
\title{Returns an array of randomly generated values from the hypergeometric distribution.}
\usage{
Rhyper(dim, m, n, k, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the hypergeometric distribution.
}
\examples{
\dontrun{}

}
