% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlnorm.R
\name{Rlnorm}
\alias{Rlnorm}
\title{Returns an array of randomly generated values from the lognormal distribution.}
\usage{
Rlnorm(dim, meanlog = 0, sdlog = 1, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the lognormal distribution.
}
\examples{
\dontrun{}

}
