% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlogis.R
\name{Rlogis}
\alias{Rlogis}
\title{Returns an array of randomly generated values from the logistic distribution.}
\usage{
Rlogis(dim, location = 0, scale = 1, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the logistic distribution.
}
\examples{
\dontrun{}

}
