% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.whsp.R
\name{rm.whsp}
\alias{rm.whsp}
\title{Extracts a subset of 3D data given by 'data', according to the array subset 'subset' or the cartesian subset 'range'.}
\usage{
rm.whsp(x, keep.lsp = FALSE)
}
\value{

}
\description{
Extracts a subset of 3D data given by 'data', according to the array subset 'subset' or the cartesian subset 'range'.
}
\examples{
\dontrun{}

}
