% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnbinom.R
\name{Rnbinom}
\alias{Rnbinom}
\title{Returns an array of randomly generated values from the negative binomial distribution.}
\usage{
Rnbinom(dim, size, prob, mu, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the negative binomial distribution.
}
\examples{
\dontrun{}

}
