% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnorm.R
\name{Rnorm}
\alias{Rnorm}
\title{Returns an array of randomly generated values from the Gaussian distribution.}
\usage{
Rnorm(dim, mean = 0, sd = 1, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the Gaussian distribution.
}
\examples{
\dontrun{}

}
