% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpois.R
\name{Rpois}
\alias{Rpois}
\title{Returns an array of randomly generated values from the Poisson distribution.}
\usage{
Rpois(dim, lambda, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the Poisson distribution.
}
\examples{
\dontrun{}

}
