% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rt.R
\name{Rt}
\alias{Rt}
\title{Returns an array of randomly generated values from the student t distribution.}
\usage{
Rt(dim, df, ncp, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the student t distribution.
}
\examples{
\dontrun{}

}
