% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runif.R
\name{Runif}
\alias{Runif}
\title{Returns an array of randomly generated values from the uniform distribution.}
\usage{
Runif(dim, min = 0, max = 1, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the uniform distribution.
}
\examples{
\dontrun{}

}
