% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rweibull.R
\name{Rweibull}
\alias{Rweibull}
\title{Returns an array of randomly generated values from the Weibull distribution.}
\usage{
Rweibull(dim, shape, scale = 1, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the Weibull distribution.
}
\examples{
\dontrun{}

}
