% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rwilcox.R
\name{Rwilcox}
\alias{Rwilcox}
\title{Returns an array of randomly generated values from the Wilcoxon distribution.}
\usage{
Rwilcox(dim, m, n, byrow = FALSE)
}
\value{

}
\description{
Returns an array of randomly generated values from the Wilcoxon distribution.
}
\examples{
\dontrun{}

}
