% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzip.R
\name{rzip}
\alias{rzip}
\title{Zips the rows of matrices into each other so that for two matrices matrix1 and matrix2 with dimensions [4,3] and [2,5] the end result is cbind(matrix1[1,],matrix2[1,],matrix1[2,],matrix2[2,],matrix1[3,],matrix1[4,]), with NAs on the spare postitions in the matrix.}
\usage{
rzip(...)
}
\value{

}
\description{
Zips the rows of matrices into each other so that for two matrices matrix1 and matrix2 with dimensions [4,3] and [2,5] the end result is cbind(matrix1[1,],matrix2[1,],matrix1[2,],matrix2[2,],matrix1[3,],matrix1[4,]), with NAs on the spare postitions in the matrix.
}
\examples{
\dontrun{}

}
