% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortby.R
\name{sortby}
\alias{sortby}
\title{Orders the columns given of the matrix 'x' by the columns given by 'by'. This function is the same as order(col1,col2,col3,...) only with a more robust argument list, where the matrix 'x' and the order 'by' in which columns are ordered are given.}
\usage{
sortby(x, by = 1, return.order = FALSE)
}
\value{

}
\description{
Orders the columns given of the matrix 'x' by the columns given by 'by'. This function is the same as order(col1,col2,col3,...) only with a more robust argument list, where the matrix 'x' and the order 'by' in which columns are ordered are given.
}
\examples{
\dontrun{}

}
