% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SX90.getSchools.R
\name{SX90.getSchools}
\alias{SX90.getSchools}
\title{***DECRIPTION MISSING***}
\usage{
SX90.getSchools(data, Ttha, includeEdge = FALSE)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
