% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeclick.R
\name{timeclick}
\alias{timeclick}
\title{***DECRIPTION MISSING***}
\usage{
timeclick(names, T = 300, pre = 5, sps = 5, score = cbind(0, 0))
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
