% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournamentPoints_2014.R
\name{tournamentPoints_2014}
\alias{tournamentPoints_2014}
\title{***DECRIPTION MISSING***}
\usage{
tournamentPoints_2014(players, tournament, date, series, pow = c(3, 3, 3, 3,
  1, 1), w = c(32, 32, 31, 28, 10, 6), bronze = TRUE, A = 1000,
  WR.out = FALSE, euro = FALSE)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
