% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournamentPointsBest.R
\name{tournamentPointsBest}
\alias{tournamentPointsBest}
\title{***DECRIPTION MISSING***}
\usage{
tournamentPointsBest(Best, level = 5, pow = c(3, 3, 3, 3, 1, 1), w = c(32,
  32, 31, 28, 10, 6), bronze = TRUE, A = 1000, players = NULL)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
