% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undrop.R
\name{undrop}
\alias{undrop}
\title{Adds dimensions to the input element that has length > 1, according to the other (1 or 2) input elements. 
NA}
\usage{
undrop(...)
}
\value{

}
\description{
Adds dimensions to the input element that has length > 1, according to the other (1 or 2) input elements. 
NA
}
\examples{
\dontrun{}

}
