% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.condexp.R
\name{var.condexp}
\alias{var.condexp}
\title{Variance of an exponential distribution restricted by X<C or X>C.}
\usage{
var.condexp(beta, C, type = "l")
}
\value{

}
\description{
Variance of an exponential distribution restricted by X<C or X>C.
}
\examples{
\dontrun{}

}
