% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbsc2informationProfile.R
\name{vbsc2informationProfile}
\alias{vbsc2informationProfile}
\title{Extracts a subset of TSD data according to the array subset 'ind' and/or the cartesian subset 'range' and/or the logical/numeric vector of subscripts 'subset'.}
\usage{
vbsc2informationProfile(event = 1, t = 1, N = 1e+05, acca = NULL,
  var = c("vbsc", "sgsc", "pr0s", "sgs0"), ind = list(-(1:100), NULL),
  range = list(), subset = NULL, cruise = 2009116, origin = c(0, 0, 0),
  angle = 0, w = 10, dens = 10, breaks = seq(-100, 100, dens), ...)
}
\arguments{
\item{ind}{is a list of indexes, as typed into the [] of an array, where 0 and NULL denotes all indexes.}

\item{range}{is a list of elements with names matching names if 'data', specifying the range of the corresponding elements.}

\item{subset}{is a numeric or logical vector/expression indicating elements or rows to keep (as used in subset()). Missing values are taken as false, and subset=0 or subset=NULL indicates no subsetting.}

\item{data}{is a list of elements named according to the TSD file format.}

\item{strict}{is TRUE if strict inequality is to be used when subsetting according to 'range'.}
}
\value{

}
\description{
Extracts a subset of TSD data according to the array subset 'ind' and/or the cartesian subset 'range' and/or the logical/numeric vector of subscripts 'subset'.
}
\examples{
\dontrun{}

}
