% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WR2linesep.R
\name{WR2linesep}
\alias{WR2linesep}
\title{Read the tournaments as saved from copy paste from the WR, and save to a file with the following structure:
NA
Tournament name
Date
Series
Number of players in A
Player_1
Player_2
...
Player_k}
\usage{
WR2linesep(dir, dir2 = NULL)
}
\value{

}
\description{
Read the tournaments as saved from copy paste from the WR, and save to a file with the following structure:
NA
Tournament name
Date
Series
Number of players in A
Player_1
Player_2
...
Player_k
}
\examples{
\dontrun{}

}
