% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePointsRound.R
\name{writePointsRound}
\alias{writePointsRound}
\title{***DECRIPTION MISSING***}
\usage{
writePointsRound(round = 1, volume = 1, pow = c(3, 3, 3, 3, 1, 1),
  w = c(32, 32, 31, 28, 10, 6), subst = TRUE,
  write = c("StatsPlayersRound", "PointsRound", "TableRound", "form"))
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
