% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTournaments.R
\name{writeTournaments}
\alias{writeTournaments}
\title{Writes all tournaments in the interval 't' (given as a two element vector c(yyyymmdd,yyyymmdd)) to one excel file.}
\usage{
writeTournaments(t = "all", file = NULL)
}
\value{

}
\description{
Writes all tournaments in the interval 't' (given as a two element vector c(yyyymmdd,yyyymmdd)) to one excel file.
}
\examples{
\dontrun{}

}
