% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTtha.R
\name{writeTtha}
\alias{writeTtha}
\title{(Internal) Calculates and writes the total horizontal area to file.}
\usage{
writeTtha(event, R = 600)
}
\value{

}
\description{
(Internal) Calculates and writes the total horizontal area to file.
}
\examples{
\dontrun{}

}
