% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWRfile.R
\name{writeWRfile}
\alias{writeWRfile}
\title{***DECRIPTION MISSING***}
\usage{
writeWRfile(players, tournament, date, series, A = NULL, bronze = TRUE,
  info = list(), euro = FALSE)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
