% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWRfile_2014.R
\name{writeWRfile_2014}
\alias{writeWRfile_2014}
\title{***DECRIPTION MISSING***}
\usage{
writeWRfile_2014(players, tournament, date, series, A = NULL, bronze = TRUE,
  info = list(), dir = NULL, euro = FALSE)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
